+++ src/quattro/_taskgroup.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import sys
+from asyncio import Semaphore
 from contextvars import Context
 from typing import TYPE_CHECKING, TypeVar
 
@@ -20,9 +21,35 @@ T = TypeVar("T")
 
 
 class TaskGroup(_TaskGroup):
-    def __init__(self) -> None:
+    def __init__(self, *, concurrency_limit: int | None = None) -> None:
+        """
+        Args:
+            concurrency_limit: When provided, use a semaphore to limit the number of
+                non-background tasks that run in parallel.
+
+        .. versionchanged:: NEXT
+           Added the `concurrency_limit` parameter.
+        """
         _TaskGroup.__init__(self)
         self._bg_tasks: set[Task] = set()
+        if concurrency_limit is not None and concurrency_limit < 1:
+            raise ValueError("concurrency_limit must be >= 1")
+        self._semaphore = (
+            None if concurrency_limit is None else Semaphore(concurrency_limit)
+        )
+
+    def create_task(
+        self,
+        coro: _CoroutineLike[T],
+        *,
+        name: str | None = None,
+        context: Context | None = None,
+    ) -> Task[T]:
+        return super().create_task(
+            coro if self._semaphore is None else _wrap_coro(coro, self._semaphore),
+            name=name,
+            context=context,
+        )
 
     def create_background_task(
         self,
@@ -37,8 +64,10 @@ class TaskGroup(_TaskGroup):
 
         If this task finishes with an error, the entire task group will be
         cancelled, like with non-background tasks.
+
+        Background tasks do not count against the concurrency limit.
         """
-        task = self.create_task(coro, name=name, context=context)
+        task = _TaskGroup.create_task(self, coro, name=name, context=context)
         if not task.done():
             self._bg_tasks.add(task)
             task.add_done_callback(lambda t: self._bg_tasks.discard(t))
@@ -61,3 +90,8 @@ class TaskGroup(_TaskGroup):
             bg_task.cancel()
 
         await _TaskGroup.__aexit__(self, et, exc, tb)
+
+
+async def _wrap_coro(coro: _CoroutineLike[T], semaphore: Semaphore) -> T:
+    async with semaphore:
+        return await coro